Removed the BeerTests class from the Api/Tests section as the tests were outdated and no longer relevant, implying that the associated tests were refactored elsewhere or became obsolete due to code changes.
Added more specific error types, replacing ArgumentNullException with more contextual exceptions like InvalidBreweryNameException, InvalidBeerNameException and InvalidBase64FormatException. This provides more context about the nature of the error to the caller.

A ValidationCommandHandlerDecorator class is added in Pijze.Infrastructure, allowing the validation of commands using FluentValidation, which provides a cleaner and more robust way to validate inputs.

Also, Beer tests are renamed and updated to Brewery tests which was more suitable in the current codebase.

Additionally, the solution file is updated to include the 'Pijze.Api.UnitTests' project.