Update Angular packages to version 15.2.9
Updated the versions of Angular dependencies to 15.2.9 for Angular cdk and material in the project's package-lock.json file. This update is in order to remain consistent with other Angular package versions in the project and to benefit from the latest updates and bug fixes from the Angular team. Also, updated the 'parse5' version and dependencies related to Angular material in line with the