Update Angular, NgRx and other dependencies versions
Updated the versions of Angular, NgRx and multiple other dependencies in package-lock.json. This change was necessary make the application current with the most recent releases, to incorporate latest updates and improvements, as well as to ensure compatibility with other recent dependency updates. These updates involve changes in multiple files to adapt to the new versions.